Need to test with ch340 chips and devise a method to get the right com port. I have selection boxes for UNO, NANO and MEGA. They work but they override the fqbn if an original arduino is plugged in. If an original arduino is plugged in, functionBlox should discard those boxes
A com port text and dummy number is added. But it is not yet in use. Method is needed to list all  com ports found by arduino-cli.

it would be neat if an event can be triggered when an arduino is plugged in

Also added name of program and me logo