relocated board files 'elsewhere'.
Finally panning works.. mostly. I still have to do the link locations.

As is now I added limits to drawing the function blocks. But I'd rather redo the background drawing in 2 parts. Not yet sure.

The links and nodes need to return the offset positions like the function blocks do.